Allow config-level setting of UAHF defaults
So when the user doesn't pass any options for the uahfstarttime the
config-option defines if it is turned on or off.
In either case, the user can override the behavior by passing a
start time. (start time of zero turns uahf chain-rules off)

Also change the seeds when UAHF is user-enabled.