Deprecate -uahfstarttime and introduce -uahf=true
We no longer allow the start time to be configurable. Users may just end
up hurting themselves as its a network-wide option.
It is deprecated, but for backwards compatibility we now interpret it as
"true" the moment its greater than zero.

Instead we add a -uahf=[true|false] option that simply enables the chain
or disbles it.

We continue to work together with the -enable-uahf config option which
turns on uahf by default.