Add mem-mapped backed Block/Transaction classes
This is an import from the validation branch,
it adds partial support for memory-mapped blocks.

This reflects the reality that as soon as a transaction is
signed, it will never change again. Making it irrelevant to parse it
into memory.
The old code reads it into structs and vectors and writes it out again
every time something needs to be done that is not possible on this
in-memory structure. For instance checking its size or calculating the
tx-id (hash).

This is worse when we look at a block, as that is a collection of
transactions.
The most absurt example is reading a block and checking if the
merkle-root is proper.

The memory-mapped solution aims to do zero copying and things like
creating a hash of all transactions in a block is indeed magnitudes
faster.