Add writing of FastBlock
this introduces the Blocks::DB::writeBlock() method
which appends the block to the last known file.

We keep the features that files are enlarged in chunks (at least on non-
win32 platforms) and we only open the last file in read/write mode so
sysadmins can move old files to a read-only filesystem.