Make initialization of Prams lazy
This cleans up the initialization from all static (before main)
to now be unique pointers owned by one static object.
This has as a neat side-effect that we only initialize the chain
we actually need, and so do less work during startup.

The main advantage is that I can use GetArg() in the constructor
since they are no longer executed before main()