Work better with misbehaving node
When we connect to a peer and send them our version message, when we
receive a version message from them we assumed they validated it to be
the same language they speak. Seems some software doesn't check the
magic version, though, which then caused us to get lots of errors in
the logs.

This change immediately disconnects a peer that replies with a version
message in the wrong encoding on outgoing connections.