Fix up address manager.
A missing feature in the address manager was that it only knew about
unreachable and reachable nodes.
The reality is that an IP that is reachable is not always useful, and we
had no way to remember a node being less than useful.

I added a simple way to remember the uselessness of a node, and if we
ban a node that score goes up. So we will try less often to connect to
it.